Tournament organisers

+++ public/officer/tournamentDelete.php
@@ -11,9 +11,20 @@ if(!isset($_GET['id'])) {
 }
 $id = $_GET['id'];
 
+$tournament = Tournament::find_by_id($id);
+
+$organiserIDs = array();
+$counter = 0;
+foreach($tournament->get_organisers() as $organiser){
+	$organiserIDs[$counter] = $organiser->id;
+	$counter++;
+}
+
+if(!in_array(get_session_id(), $organiserIDs)){
+	redirect_to(url_for('officer/tournaments.php?id=' . $id));
+}
+
 if(is_post_request()) {
-	
-	$tournament = Tournament::find_by_id($id);
 	$result = $tournament->delete();
 	
 	if($result){
